LPS-114993 Organize components
