+++ modules/apps/layout/layout-seo-web/src/main/resources/META-INF/resources/js/seo/display_page_templates/SeoMapping.js
@@ -14,7 +14,7 @@
 
 import React from 'react';
 
-import MappingField from './MappingField';
+import MappingField from './components/MappingField';
 
 function SeoMapping({
 	description,
