+++ modules/apps/layout/layout-seo-web/src/main/resources/META-INF/resources/js/seo/display_page_templates/components/MappingField.js
